move atac fixture setup to valid_adatas and update some 5.3 atac tests
